changed the logic, changed the links on the site

+++ src/main/java/com/example/demo/job/ParseTour.java
@@ -11,6 +11,7 @@ import org.openqa.selenium.*;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.remote.SessionId;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
@@ -52,20 +53,11 @@ public class ParseTour {
     int counterForTEst = 0;
 
     //    @Scheduled(fixedDelay = 25_000_000)
-    @Scheduled(fixedDelay = 10_000)
+    @Scheduled(fixedDelay = 18_000_000)
     public void startParse() {
+        SessionId sessionId;
         counterForTEst++;
-        ChromeOptions options = new ChromeOptions();
-
-        // load page Strategy
-        options.setPageLoadStrategy(PageLoadStrategy.EAGER);
-        // hide browser
-        options.addArguments("--headless");
-        options.addArguments("--no-sandbox");
-        options.addArguments("--blink-settings=imagesEnabled=false");
-        options.addArguments("--incognito");
-        options.addArguments("--disable-gpu");
-        WebDriver webDriver = null;
+        ChromeDriver webDriver = null;
 
         if (counterForTEst == 1) {
             FunSunRestController.parsLinks.add("https://fstravel.com/searchtour/country/africa/egypt?departureCityId=244707&arrivalCountryId=18498&minStartDate=2024-08-01&maxStartDate=2024-08-03&minNightsCount=10&maxNightsCount=14&adults=2&flightTypes=charter&sort=max&mealTypes=10004,8&distanceSea=04093636-6324-4c8e-878b-a27871420711&regions=18535&cities=117164,21094,122878,117154,117163&stars=4,5");
@@ -73,7 +65,6 @@ public class ParseTour {
             FunSunRestController.parsLinks.add("https://fstravel.com/searchtour/country/africa/egypt?departureCityId=244707&arrivalCountryId=18498&minStartDate=2024-08-07&maxStartDate=2024-08-07&minNightsCount=10&maxNightsCount=14&adults=2&flightTypes=charter&sort=max&mealTypes=10004,8&distanceSea=04093636-6324-4c8e-878b-a27871420711&regions=18535&cities=117164,21094,122878,117154,117163&stars=4,5");
             FunSunRestController.parsLinks.add("https://fstravel.com/searchtour/country/africa/egypt?departureCityId=244707&arrivalCountryId=18498&minStartDate=2024-08-08&maxStartDate=2024-08-08&minNightsCount=10&maxNightsCount=14&adults=2&flightTypes=charter&sort=max&mealTypes=10004,8&distanceSea=04093636-6324-4c8e-878b-a27871420711&regions=18535&cities=117164,21094,122878,117154,117163&stars=4,5");
             FunSunRestController.parsLinks.add("https://fstravel.com/searchtour/country/africa/egypt?departureCityId=244707&arrivalCountryId=18498&minStartDate=2024-08-09&maxStartDate=2024-08-11&minNightsCount=10&maxNightsCount=14&adults=2&flightTypes=charter&sort=max&mealTypes=10004,8&distanceSea=04093636-6324-4c8e-878b-a27871420711&regions=18535&cities=117164,21094,122878,117154,117163&stars=4,5");
-
             FunSunRestController.parsLinks.add("https://fstravel.com/searchtour/country/africa/egypt?departureCityId=244707&arrivalCountryId=18498&minStartDate=2024-08-13&maxStartDate=2024-08-14&minNightsCount=10&maxNightsCount=14&adults=2&flightTypes=charter&sort=max&mealTypes=10004,8&distanceSea=04093636-6324-4c8e-878b-a27871420711&regions=18535&cities=117164,21094,122878,117154,117163&stars=4,5");
             FunSunRestController.parsLinks.add("https://fstravel.com/searchtour/country/africa/egypt?departureCityId=244707&arrivalCountryId=18498&minStartDate=2024-08-15&maxStartDate=2024-08-15&minNightsCount=10&maxNightsCount=14&adults=2&flightTypes=charter&sort=max&mealTypes=10004,8&distanceSea=04093636-6324-4c8e-878b-a27871420711&regions=18535&cities=117164,21094,122878,117154,117163&stars=4,5");
             FunSunRestController.parsLinks.add("https://fstravel.com/searchtour/country/africa/egypt?departureCityId=244707&arrivalCountryId=18498&minStartDate=2024-08-16&maxStartDate=2024-08-17&minNightsCount=10&maxNightsCount=14&adults=2&flightTypes=charter&sort=max&mealTypes=10004,8&distanceSea=04093636-6324-4c8e-878b-a27871420711&regions=18535&cities=117164,21094,122878,117154,117163&stars=4,5");
@@ -89,33 +80,67 @@ public class ParseTour {
         try {
             if (!FunSunRestController.parsLinks.isEmpty()) {
                 for (int i = 0; i < FunSunRestController.parsLinks.size(); i++) {
-                    webDriver = new ChromeDriver(options);
+                    System.out.println();
+                    System.out.println();
+                    System.out.println();
+                    // check for close browser
+                    if (webDriver != null) {
+                        System.out.println("not null");
+                        webDriver.quit();
+                        webDriver = null;
+                    }
+                    webDriver = new ChromeDriver(getChromeOptions());
+                    sessionId = webDriver.getSessionId();
                     totalErrorCount = 0;
                     countForRemoveLink = 0;
                     idLinkInWork = i;
                     webDriver.get(FunSunRestController.parsLinks.get(i));
                     sleep();
-
+                    System.out.println("Array " + i);
                     try {
-                        for (int j = 1; j < 15; j++) {
+                        for (int j = 3; j < 15; j++) {
                             funSunParse(webDriver, j);
                         }
                     } catch (NoSuchSessionException ignore) {
+                        System.out.println(126);
                     }
-
                     try {
-                        System.out.println("webDriver.quit() Array = " + i);
-                        webDriver.manage().deleteAllCookies();
-                        webDriver.quit();
+                        if (webDriver.getSessionId() == sessionId) {
+                            System.out.println("webDriver.quit()");
+                            System.out.println("===========" + i);
+                            System.out.println();
+                            System.out.println();
+                            System.out.println();
+                            System.out.println();
+                            System.out.println();
+                            webDriver.manage().deleteAllCookies();
+                            webDriver.quit();
+                            webDriver = null;
+                        } else {
+                            System.out.println("==========");
+                            System.out.println("==========");
+                            System.out.println("==========");
+                            System.out.println("==========");
+                            System.out.println("==========");
+                            System.out.println("==========");
+                            System.out.println("==========");
+                            System.out.println("==========");
+                            System.out.println("==========");
+                            System.out.println("==========");
+                            System.out.println("==========");
+                        }
                     } catch (TimeoutException e) {
-                        System.out.println("timeout 108");
-                        webDriver.navigate().refresh();
-                        System.out.println("timeout 108");
+                        System.out.println(135);
+                        e.printStackTrace();
+                        System.out.println(("Page load Timeout Occured. Quiting !!! 184"));
+                    } catch (NoSuchWindowException e) {
+                        return;
                     }
                 }
 
             }
         } catch (ConcurrentModificationException e) {
+            System.out.println(142);
             errorLog("Mod array", webDriver);
             return;
         }
@@ -140,19 +165,32 @@ public class ParseTour {
                 currentUrl = webDriver.getCurrentUrl();
                 webDriver.close();
                 webDriver.switchTo().window(arrayWithTab.get(0));
-                arrayWithTab.clear();
+                arrayWithTab = null;
             } catch (NoSuchElementException | ElementClickInterceptedException e) {
+                System.out.println(171);
+                return;
+            } catch (WebDriverException e) {
                 return;
             }
 
             String rating;
             String price;
             String reviews;
+            String flyDate;
 
             hotelName = webDriver.findElement(By.cssSelector("#app > div > div.v-main > div.container > div > div > div.search-content > div.search-content__cards.tour-search-content__result > div.search-cards-container > div:nth-child(" + hotel + ") > div.hotelCard__description > div.hotelCard__description-header > div.hotelCard__description-header_text > div.hotelCard__description-header_text-name")).getText();
             location = webDriver.findElement(By.cssSelector("#app > div > div.v-main > div.container > div > div > div.search-content > div.search-content__cards.tour-search-content__result > div.search-cards-container > div:nth-child(" + hotel + ") > div.hotelCard__description > div.hotelCard__description-header > div.hotelCard__description-header_text > div.hotelCard__description-header_text-location")).getText();
+
             flyDate = webDriver.findElement(By.cssSelector("#app > div > div.v-main > div.container > div > div > div.search-content > div.search-content__cards.tour-search-content__result > div.search-cards-container > div:nth-child(" + hotel + ") > div.hotelCard__description > div.hotelCard__description-flight_price > div.hotelCard__description-flight_price-definition > div.hotelCard__description-flight_price-definition_information")).getText();
-            countNight = webDriver.findElement(By.cssSelector("#app > div > div.v-main > div.container > div > div > div.search-content > div.search-content__cards.tour-search-content__result > div.search-cards-container > div:nth-child(" + hotel + ") > div.hotelCard__description > div.hotelCard__description-flight_price > div.hotelCard__description-flight_price-definition > div.hotelCard__description-flight_price-definition_information")).getText();
+
+            if (flyDate.length() == 30) {
+                this.flyDate = flyDate.substring(0, 5);
+                this.countNight = flyDate.substring(22, 24);
+            }
+            if (flyDate.length() == 31) {
+                this.flyDate = flyDate.substring(0, 6);
+                this.countNight = flyDate.substring(23, 25);
+            }
 
 
             try {
@@ -163,33 +201,48 @@ public class ParseTour {
                 this.reviews = Integer.parseInt(reviews.replaceAll("[^\\d.]", ""));
                 this.rating = Float.parseFloat(rating.replaceAll("[^\\d.]", ""));
             } catch (NumberFormatException | NoSuchElementException e) {
+                System.out.println(194);
                 this.reviews = 0;
                 this.rating = 0;
             }
 
         } catch (org.openqa.selenium.NoSuchElementException e) {
+            System.out.println(200);
             errorLog("NoSuchElementException", webDriver);
         } catch (ElementClickInterceptedException e) {
+            System.out.println(203);
             errorLog("ClickInterceptedException", webDriver);
         } catch (JavascriptException exception) {
+            System.out.println(206);
             errorLog("JavascriptException", webDriver);
-        } catch (TimeoutException e) {
-            System.out.println("timeout 169");
-            webDriver.navigate().refresh();
-            System.out.println("timeout 169");
-        } catch (NoSuchWindowException ignore) {
-
+        } catch (NoSuchWindowException e) {
+            System.out.println(209);
+        } catch (org.openqa.selenium.TimeoutException e) {
+            System.out.println(210);
+            e.printStackTrace();
+            System.out.println(("Page load Timeout Occured. Quiting !!! 184"));
         }
+
+
         // no such element
         if (currentUrl != null) {
             if (this.price <= 220000 && this.rating >= 4) {
-
-                System.out.print(hotelName);
-                System.out.println(" " + this.price);
-
                 workWithDataBase();
             }
         }
+
+        hotelName = null;
+        price = 0;
+        location = null;
+        flyDate = null;
+        countNight = null;
+        rating = 0;
+        reviews = 0;
+        currentUrl = null;
+    }
+
+    private boolean isEmpty(WebDriver webDriver, String cssString) {
+        return webDriver.findElement(By.cssSelector(cssString)).getText().isEmpty();
     }
 
     private void workWithDataBase() {
@@ -203,14 +256,18 @@ public class ParseTour {
             tourInfoEntity = new TourInfoEntity(hotelName, price, location, flyDate, countNight, rating, reviews, currentUrl, 0);
             tourService.saveTour(tourInfoEntity);
         }
-        // update tour price
+        // update tour
         if (tourInfoRepo.findByURL(currentUrl).getTourPrice() != price) {
-            tourInfoRepo.findByURL(currentUrl).setDifferenceInPrice(tourInfoRepo.findByURL(currentUrl).getTourPrice() - price);
+//            tourInfoRepo.findByURL(currentUrl).setDifferenceInPrice(tourInfoRepo.findByURL(currentUrl).getTourPrice() - price);
+            System.out.println(tourInfoRepo.findByURL(currentUrl).getTourPrice());
+            System.out.println(currentUrl);
+            tourInfoRepo.updatePrice(tourInfoRepo.findByURL(currentUrl).getTourPrice() - price, currentUrl);
         }
+        tourInfoEntity = null;
     }
 
     private void sleep() {
-        int random = new Random().nextInt(6000) + 3000;
+        int random = new Random().nextInt(35000) + 20000;
         try {
             Thread.sleep(random);
         } catch (InterruptedException ex) {
@@ -219,7 +276,9 @@ public class ParseTour {
     }
 
     // Error log to DB
+
     private void errorLog(String errorCode, WebDriver webDriver) {
+        System.out.println(errorCode);
         DateFormat dateFormat = new SimpleDateFormat("dd/MM/yy HH:mm:ss");
         Date date = new Date();
 
@@ -246,6 +305,21 @@ public class ParseTour {
         countForRemoveLink++;
         totalErrorCount++;
 
+    }
+
+    // Chrome Options
+    private static ChromeOptions getChromeOptions() {
+        ChromeOptions options = new ChromeOptions();
+        // load page Strategy
+        options.setPageLoadStrategy(PageLoadStrategy.NORMAL);
+        options.addArguments("--headless");
+        options.addArguments("--no-sandbox");
+        options.addArguments("--blink-settings=imagesEnabled=false");
+        options.addArguments("--incognito");
+        options.addArguments("--disable-gpu");
+        options.addArguments("--disable-popup-blocking");
+        options.addArguments("--Disable-extensions");
+        return options;
     }
 
 }
